LayoutTests/http/wpt/service-workers/extendableEvent.https.html

<!DOCTYPE html>
<title>Service Worker Extendable Event</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
// FIXME: Should be run on a service worker.
test(() => {
    if (!window.internals)
        return Promise.reject("test require internals");
    var event = new ExtendableEvent('ExtendableEvent', {});
    assert_throws('InvalidStateError', () => event.waitUntil(new Request('')));
}, "ExtendableEvent waitUntil should support non promise parameters");

test(function() {
    assert_equals(
      new ExtendableEvent('ExtendableEvent').type,
      'ExtendableEvent', 'Type of ExtendableEvent should be ExtendableEvent');
    assert_equals(
      new ExtendableEvent('ExtendableEvent', {}).type,
      'ExtendableEvent', 'Type of ExtendableEvent should be ExtendableEvent');
    assert_equals(
      new ExtendableEvent('ExtendableEvent', {}).cancelable,
      false, 'Default ExtendableEvent.cancelable should be false');
    assert_equals(
      new ExtendableEvent('ExtendableEvent', {}).bubbles,
      false, 'Default ExtendableEvent.bubbles should be false');
    assert_equals(
      new ExtendableEvent('ExtendableEvent', {cancelable: false}).cancelable,
      false, 'ExtendableEvent.cancelable should be false');
  }, 'Event constructors');
</script>
